Add data to firestore
Some of the data has been uploaded to a firestore db. However, the
free-tier daily write limit (20k documents) is too small to complete the
entire transaction in one day (as we have ~48k documents for images
alone and another 7k for image groups). Completing this will require
additional code to identify which image groups have not yet been
uploaded.